Move PBEM 'editors' to PBEM package
