+++ game-core/src/main/java/games/strategy/engine/framework/startup/mc/SetupPanelModel.java
@@ -8,9 +8,9 @@ import games.strategy.engine.framework.startup.login.ClientLoginValidator;
 import games.strategy.engine.framework.startup.ui.ClientSetupPanel;
 import games.strategy.engine.framework.startup.ui.LocalSetupPanel;
 import games.strategy.engine.framework.startup.ui.MetaSetupPanel;
-import games.strategy.engine.framework.startup.ui.pbem.PbemSetupPanel;
 import games.strategy.engine.framework.startup.ui.ServerSetupPanel;
 import games.strategy.engine.framework.startup.ui.SetupPanel;
+import games.strategy.engine.framework.startup.ui.pbem.PbemSetupPanel;
 import games.strategy.engine.lobby.client.login.LobbyLogin;
 import java.awt.Dimension;
 import java.util.Optional;
