+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/pbem/ForumPosterEditor.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.framework.startup.ui.editors;
+package games.strategy.engine.framework.startup.ui.pbem;
 
 import java.awt.GridBagConstraints;
 import java.awt.Insets;
