+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/pbem/PbemSetupPanel.java
@@ -11,10 +11,6 @@ import games.strategy.engine.framework.startup.mc.HeadedLaunchAction;
 import games.strategy.engine.framework.startup.ui.PlayerSelectorRow;
 import games.strategy.engine.framework.startup.ui.PlayerType;
 import games.strategy.engine.framework.startup.ui.SetupPanel;
-import games.strategy.engine.framework.startup.ui.editors.DiceServerEditor;
-import games.strategy.engine.framework.startup.ui.editors.EmailSenderEditor;
-import games.strategy.engine.framework.startup.ui.editors.ForumPosterEditor;
-import games.strategy.engine.framework.startup.ui.editors.ForumPosterEditorViewModel;
 import games.strategy.engine.posted.game.pbf.IForumPoster;
 import games.strategy.engine.random.PbemDiceRoller;
 import java.awt.Dialog;
