+++ game-core/src/test/java/games/strategy/engine/framework/startup/ui/editors/ForumPosterEditorTest.java
@@ -1,6 +1,6 @@
 package games.strategy.engine.framework.startup.ui.editors;
 
-import static games.strategy.engine.framework.startup.ui.editors.ForumPosterEditor.isInt;
+import static games.strategy.engine.framework.startup.ui.pbem.ForumPosterEditor.isInt;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
